Mount a different context for preprod deployment
The deployment namespace is controlled via the `KUBE_ENV_NAMESPACE`
environment variable. To override it, we need to use a different
CircleCI context.

(Discovered via `circleci config process .circleci/config.yml`)

Also changed to the `hmpps-common-vars` context, which is now the
default for all HMPPS apps for Slack notifications and other common
things.