Add an intervention progress page
The data here is all currently hardcoded, and some of the states shown
in the designs (e.g. sessions in progress) will be built as we build
that functionality.

The designs show what looks like a hyperlink for the "create action
plan" action, but we need to use a <button> to submit the form so I’m
just going to accept the default secondary button styling.