Add contract for updating Relevant Sentence ID
Users will select the relevant sentence in the Refer journey, based
on values fetched from the Community API for that Service User's CRN.

Although SentenceID is a `long` type in the Community API, we're going
to send it as a `number` here as we think it should be safe to do so.
When we deserialize it from the Community API response, we'll raise an error
as a warning if this were to overflow the Number type.