Add RelevantSentencePresenter
To present title and information about any sentences linked to the
referral.

In the designs, we display an "Order" to the user as well as the main
offence's `category` and `subcategory`. It's not clear where this comes
from at this stage, but we've decided to leave it for the time being.

This currently doesn't play back any previously-selected values, so I've
hardcoded `checked` to `false` until we introduce that functionality.