Add a ClockTime class
This is the time equivalent of our CalendarDay class - a way of
representing the values displayed by a clock on the wall, independent of
time zone.

We also add methods for getting the time as a 12-hour clock time, since
the "time" input on the upcoming "edit action plan appointment" page
uses 12-hour time input.