Add a script to generate macro args types
We’ve been using TypeScript types to represent the arguments taken by
the different GOV.UK Frontend component Nunjucks macros. This gives us
an easy way to check the arguments that our views are passing to the
macros.

Until now we’ve been generating these TypeScript types by hand, based on
the various “Nunjucks macro options” provided for each component on the
Design System website; e.g. https://design-system.service.gov.uk/components/text-input/.

This has been an ad-hoc and slightly tedious process. I happened to
realise that the website documentation is actually generated from
macro-options.json files which are included alongside each component in
the Frontend repo. So here we add a script

> `npm run-script generate-frontend-types`

which processes those JSON files and emits a file of TypeScript
type declarations.

For now, this file needs to be manually re-generated when the GOV.UK
Frontend gets updated, and the new file checked in to version control.
Perhaps in the future we could make this an automatic part of the build
pipeline, but it's not very important.

If the script for generating this file for some reason stops working
with future versions of the Frontend, and we don’t want to spend time
fixing it, we can still keep the output file and return to manual
alterations by hand.

A possible further improvement here would be to reduce the number of
type assertions in the code, or make them safer by using some sort of
JSON Schema check of the macro-options.json files.

(The change to InteventionProgressView is to fix a type name which
changed after switching to the generated types.)