Refactor formatDateTimeOrEmptyString
by adding separate functions for extracting the date and time from a
DateTime string.

We'll be using the individual extracted functions to display just the
date and just the time in the Post Session Feedback form, so it makes
sense to return the individual components and then join them, rather
than join them and split them again.

This isn't all that DRY, but I couldn't think of a DRY way of doing this
that wasn't just as verbose.

These should actually be in 12hr format according to the designs, but
I'd like to do that later in a separate piece of work (IC-1452), as it
applies to other parts of the application.