Update script to generate types for custom component macro args
This generates TypeScript types for all of the components in
server/views/components.

I’d have liked to put these types into a separate file, instead of
putting them into the existing govukFrontendTypes file. But, since the
custom types refer to types from the GOV.UK Frontend, we’d need to
import all of those types into the second file, and that would add a bit
of extra work I don’t want to get sucked into now. So, for now, we put
everything into the same file, even if that file is a now a little bit
misnamed.