Add post session feedback confirmation presenter
This is very similar to the action plan confirmation screen, although we
pass in the PP's name here. In future, we may choose to store the PP on
the referral (when we confirm the 'responsible officer' in the Refer
journey). If that's the case, we might be able to reuse the action plan
confirmation presenter.

The designs for this screen include the date of the next session, but
this will be added in the future.