User can schedule an action plan appointment
In the designs, all of the fields are marked as optional. However,
speaking with Tom, we decided that for now it’s okay for these to be
mandatory. (It would be more implementation effort to make them
optional, and also our current API contract requires that you submit a
datetime, meaning that the user would need to set either both or neither
of date and time.)