Extract CalendarDay logic from completion deadline form
We extract it into a reusable "form input" class, which we can use in
any form that needs a calendar day input. We’ll use this on the upcoming
"edit action plan appointment" page.

This also removes the data sanitisation from the form - this is
something that we tried with earlier forms and didn’t continue since it
wasn’t particularly useful (the presenter needs to be able to handle
non-number inputs anyway, so might as well just always convert things to
strings).

It also proposes a change to the form classes’ interfaces, returning a
"form data" object which _either_ has an error or a paramsForUpdate
(which is enforced by the type system). This means we don't need to do
things like returning {} from paramsForUpdate if there was actually an
error.