Get all sessions for an action plan at once.
This pulls out some nasty code which created empty sessions for
missing appointments on action plans. This has been changed on the
backend now so that an action plan that has been submitted always
has the correct number of appointments automatically created.

There is no longer a need for a 'create appointment' endpoint so the
contracts for this have been removed.