Return a single error when there are multiple calendar day errors
This brings us closer in line with the GOV.UK Design System guidance,
which says that you should only return a single error for a date. Their
exact guidance is that you should return the "highest priority" error,
which we’re not necessarily doing here, but we can refine this later in
IC-706.