Make runValidations not have side effects
I realised that running express-validator’s validations causes it to
mutate the request to store error information (by adding a
express-validator#contexts property). This means that if you run
multiple validation chains on the same request, then subsequent results
will contain the errors from previous chains, ultimately leading to
duplicated errors. This hasn’t been a problem until now, because we only
run a single chain on each request, but it will be now that I’m going to
split the forms into multiple input validation classes, each of which
runs a chain.

The simplest way to solve this is to clone the request before passing it
to express-validator.