Add tests for Community API and pass in RestClient
We didn't have any tests for this, and I'm about to introduce some new
functionality to the Community API Service to do things with the Service
User address.

As we refactored RestClient as part of setting up the
AssessRisksAndNeedsService to taken an optional token on each HTTP
request, this has required a minimal amount of work. We'll probably want
to update this on the InterventionsService and eventually stop passing
the token to RestClient on initialization.