update ActionPlanAppointment interface to match API DTOs
this commit changes the 'nullability' of the session feedback
fields in the appointment model. sessionfeedback is always supplied,
but the values in 'attendance' and 'behaviour' themselves are nullable.
the 'submitted' field is a non-nullable boolean.

a knock on effect of this change is that the check in 'interventionProgressPresenter'
to determine if the PP can view the feedback must change from checking if
`sessionFeedback.attended` is null or undefined, to checking if `sessionFeedback.submitted`
is false.