Clarify entire appointment present in response
When developing another feature relying on these fields, I found it
difficult to determine whether the session feedback fields (attendance
and behaviour) would be present in the response from the API or not,
particularly before they'd been sent vs if they had been skipped as an
optional answer.

By ensuring we use the `actionPlanAppointmentFactory` in this step, we
know that the fields will be `null` but present on the response object
by default, as that's how they're configured in the factory.

I've also added the missing parts of the response to make sure they're
in every contract test to be totally clear what's coming back from every
request.