Remove unused static pages
The idea of these pages was for the designers to get going on building
some of the templates for the MVP service. But now, we’ve already built
the critical outcomes version of most of these pages, and we will
iterate on those instead of switching to these "MVP" versions.

This removes the custom CSS, too, which means that when we come to set
some standard for how we structure and write our CSS, we’ll be able to
do so pretty much from scratch, which is nice.

I’ve kept on example static page to keep the tests passing. I haven’t
removed the static pages mechanism because it might still prove useful
as the designers continue to work on the project, perhaps for
prototyping or demoing components.