Extract answer fetching to own presenter
The `FeedbackAnswersPresenter` can be used by the Check Answers
presenters and in the upcoming View Feedback presenters to re-use the
behaviour for displaying answers selected by the SP.

Now that we're no longer passing in the `AttendancePresenter` to the
`PostSessionAttendanceFeedbackPresenter`, I've duplicated the
`sessionDetailsSummary` function to ensure we're still displaying the
session details on that page.

I initially considered adding the sessionDetailsSummary to the
FeedbackAnswersPresenter, or including the `AttendancePresenter` again,
but as this is re-used across different user journeys, the details
displayed will change depending on where it's used, so I've opted to
give each Presenter its own version for now.