Add layout and presenter for end of service report form pages
These will be used for every page on the upcoming "create end of service
report" journey. The approach is similar to that taken by the action
plan journey, except this time we have a dynamic number of pages (since
it’s determined by the number of outcomes in the referral).