Probation practitioner can view submitted end of service report
This mainly reuses the service provider "check your answers" page, with
a different page layout, some introductory text, and without the
"Change" links.

It is currently only accessible by URL (which will be included in the
notification email sent to PPs when an SP submits and end of service
report). In the future I imagine we’ll want to link to it from the PP
intervention progress page.