Add new 'generic' ReferralDTO class and GET /referral/{id} endpoint.
The new DTO combines the existing draft, sent and ended referral DTOs.
The DTO has several top level fields like id, createdAt, serviceCategoryId
that are independent of the 'state' of the referral. The rest of fields
are split between nullable nested objects that represent fields relating
to draft, sent and ended referrals respectively.