Rename ServiceCategory type to ServiceCategoryFull.
This is the precursor to some bigger changes where the referral object
contains the service category name instead of just the ID. In that case
the ServiceCategory model will be smaller, not containing the long lists
of complexity levels or outcomes.

For the time being I've just renamed this existing type with no other
changes to ensure it's clear this is the 'full' version of the model.