Add new 'getReferral' method to intervention service.
This utilizes the new API endpoint that returns a 'generic'
Referral DTO which can represent either a draft, sent, or ended
referral. The service category description is also included which
simplifies to client code here by removing the need for an extra API
call.

The only place this is currently used is in the SP referral progress
page, which is mainly illustrative. There are many other places where
we will want to change to using this endpoint in the future.

This commit does not include any contract tests. These will come later.