Redirect to check-answers page on form submit
This feels a bit hacky at the moment because we're not storing a draft
"Cancellation" on the backend, we're passing the values between pages
in-session.

Because express doesn't allow you to pass params to a redirect
controller action, I've had to make this a `render`, but it doesn't feel
quite right.

This commit also adds a Form object to prevent the Referral Cancellation
form from being submitted if a reason for cancelling the referral is not
selected.