Add missing CSRF token
This has caught me out a couple of times, when the backend hasn't yet
been ready so I've been relying on the integration tests to test a
feature. The CSRF checks are disabled for the tests, which makes sense
maybe for the unit tests, but I think not for integration. We should
consider enabling them for the integration tests, but here I’m just
fixing the immediate issue.