Make layout presenter / view able to populate service user banner
To give a template the serviceUserNotificationBannerArgs that it needs
to render the service user banner, a controller now just needs to pass a
service user to ControllerUtils.renderWithLayout.

We have some existing code to clean up in the places where we have
individual presenters and views still doing this setup by themselves.
I’ll do this in the next commit.