Add service user banner to end of service report pages
I deliberately left the service user banner out when I built these pages
originally, because I didn’t want to contribute to the growing pile of
boilerplate that we’d been using to handle the banners. Now that, as of
a7e128f, it’s a lot easier to display the SU banner, we can now add
them.

There are other places in the codebase where we should have the SU
banner but don’t. We should come back and add them there at some point,
too, but I want to keep the current batch of work relatively focused.