Break Session Table HTML down into structured data
Passing structured data around the application rather than long HTML
strings feels safer and less brittle to test, as we're not testing whole
strings and asserting that the HTML attributes are in a specific order.
It should be the View's responsibility to add HTML to the data passed to
it.

Add ProgressSessionTableRow to prevent compilation errors when trying to
access new fields in the view on `unknown` values in the previous
`Record<string,unknown>