Make link text & href nullable rather than strings
Previously, we were using an empty string when we didn't want to display
a link in the session progress table, and checking whether the string
was empty when rendering the link.

Making it nullable makes it clearer what we expect here, and is more
consistent with our nullable pattern throughout the codebase.