Refactor all user / auth related types into the models directory.
This set of changes takes all the various user types and associated models
and tidies them up into a 'models/hmppsAuth' directory.

Part of this change involves removal of the 'userService' class, since
it was really doing hmpps auth stuff, and removal of some name formatting
logic from a method that really shouldn't be doing it. The new methods in
its place do what they say on the tin, and the calling code ('passport.ts')
wranges the name into the format it wants.

move AuthUser to the hmpps models dir and rename it to 'User'