Add serviceCategoryIds property to referral
We’re adding this to support “cohort” interventions - that is,
interventions that support a range of service categories, a subset of
which the referring probation practitioner selects in the referral.

This property will eventually replace the existing `serviceCategoryId`
property, but to avoid having to make a big-bang change across the
codebase we’ll keep both initially and remove the old property later.