An intervention can have multiple service categories
As in 206c3f4, we’re adding this to support cohort interventions. An
intervention’s serviceCategories represents the possible list of service
categories that a probation practitioner can choose from when making a
referral for that intervention.

Also as in 206c3f4, we’re initially adding support for multiple service
categories alongside the existing single service category, and then in
IC-1594 we’ll migrate the existing single category code and then get rid
of the serviceCategory property.

This also fixes an issue with the original implementation that didn't
correctly access the `sequence` property, meaning calling `toString` on
it was returning an object without us realising.