Add interventionId to referral types
When a probation practitioner is making a referral to a cohort
intervention, we will present them with a list of the possible service
categories allowed by that intervention, and they will make a selection
from this list.

To fetch this list, we need to know which intervention the referral is
being made against. After fetching the intervention, we can then use the
serviceCategories property added in 3b2a66a to fetch the service
categories.