Allow setting serviceCategoryIds on a draft referral
Until now, the service category was set automatically by the API as part
of creating the draft referral, using the single service category in the
intervention. But this won’t work for cohort interventions, where the
probation practitioner needs to make a selection from a list of service
categories allowed by the intervention.

(We’ll have to figure out whether this will have any effect on
non-cohort interventions — will the frontend need to start explicitly
setting the service category for those ones, too, or will the API
continue to handle that upon draft referral creation?)