Add select service categories page
This is used to set the service categories on a cohort referral. It's
currently not linked to by the task list, as we don't want to break
the existing journey for non-cohort referrals.

In future, we'll render a link to this when there's a referral for an
intervention with multiple service categories on it.