Add CohortDesiredOutcomes to draft referral
and contract test for accessing these values.

This will store the user-selected desired outcomes on a cohort referral
against the relevant service category.

I'm not 100% sure about the naming of this, as it may be the case that
we want this to be a total replacement for how we're currently storing
`desiredOutcomesIds` for single-service referrals, in which case, we can
probably drop the `cohort` but I thought I'd leave it on there for now
to make it backwards compatible, but we may want to just call it
something like `desiredOutcomes` in future.