Add ReferralDesiredOutcomes to draft referral
and contract test for accessing these values.

This will store the user-selected desired outcomes on a cohort referral
against the relevant service category.

This is necessary as we currently make a call `patchDraftReferral` with
the desiredOutcomesIds on one page, then on the next page (for the next
service category) we want to add to the `desiredOutcomesIds` array
without overwriting the previously-set ids. I initially thought we could
just fetch the previous answers and re-set them all in one go, but if a
user goes back to change their answers (e.g. and unselects a box), we
want to make sure we don't overwrite all desired outcomes for other
service categories.

Eventually,  we want this to be a total replacement for how we're
currently storing `desiredOutcomesIds` for single-service referrals, so
I've marked that field as deprecated to clarify that it should be
removed in future..