Add `viewCohortDesiredOutcomes` controller action
This uses existing presenters and views but is built to handle referrals
with an array of `serviceCategoryIds`. It's not currently linked up to
the referral form, that will come in a future story.

I've added this, rather than replaced the original `viewDesiredOutcomes`
route, as I don't want to break existing functionality, but we'll
probably be able to use this action for both cohort and single-service
referrals once the backend has been updated.