Set desired outcomes on cohort referrals
This updates the method name to combine viewing and updating desired
outcomes in the controller via a conditional check on whether it's a
PATCH request or not.

This commit temporarily switches the type of `desiredOutcomes` on
`setDesiredOutcomesForServiceCategory` to be a `Partial<DraftReferral>`
rather than `Partial<ReferralDesiredOutcomes>` so that we can use the
form's response in its current implementation (shared between the two
different ways of setting desired outcomes ids, either on single-service
and cohort referrals). Once we switch the single-service referrals to
set desired outcomes in this way, we can change the type back on that
argument.