Add setDesiredOutcomesForServiceCategory
This is used to set desired outcomes for one service category, rather
than the usual patch request we make that sets all desired outcomes ids
at once on the referral.

This is necessary for cohort referrals where we want to set the desired
outcomes ids individually, as we want to be able to overwrite them when
a user comes back to the relevant screen e.g. to change their answers.

I've opted for a PATCH, rather than a PUT here, as PATCH should be used
for a partial update to a resource, while PUT should be used when
replacing that resource. Although we are effectively replacing the
`desiredOutcomes` resource, that resource is actually on the `Referral`,
which is the entity we're updating.

This will eventually replace the call to `patchDraftReferral` everywhere
we set desired outcomes.