Remove extra desired outcome in sentReferral test
This was added in the previous branch that added cohort referrals, but
it's worth adding a specific test case for those in future, rather than
tacking them on here.

This is more for consistency now that we're about to add a new
`complexityLevel` field, and only want to add it for a single-service
referral here.