Add missing test for referral complexity level id
It looks like we didn't have one of these asserting that a complexity
level id exists on a referral. It feels like we should have one to
mirror the one we have for desired outcomes.

In future, we'll be able to remove/update this once we switch over to
the new structure for both single and cohort referrals.