Add ReferralComplexityLevel to draft referral
and contract test for viewing these values on a cohort referral.

This change allows us to associate a complexity level with a given
service category, like we do with Desired Outcomes, as now that cohort
referrals have been brought into the mix, there's the possibility of a
referral having multiple service categories selected on it.

We'll also be able to add a new endpoint for PATCHing a single
complexity level on a draft referral, like we do with Desired Outcomes.