Add setComplexityLevelForServiceCategory
This will replace the existing generic PATCH request we make to set the
Complexity Level on a referral, now that we have cohort referrals which
might have multiple Service Categories.

Although it's specifically designed for cohort referrals, it will work
with single-service referrals too, so we should refactor existing
behaviour to use this new function once it's been implemented.