Refactor ComplexityLevelForm to new pattern
We've now moved towards using a `data` function to return the params to
be updated or an error in other parts of the service, and relying on
"express-validator" to handle our validations.

This came about as I want to reuse the same form for selecting
complexity ids on a cohort referral.