refactor ReferralAccessChecker so it throws AccessErrors instead of returning a Boolean.
the motivation behind this change is twofold. firstly it means that the caller does not have
to be explicit about checking the return value and then dealing with revoking access itself;
if access is forbidden, an AccessError is thrown immediately. secondly, it means this class
can provide further information about the authorization issues in the same way we do in the
scope mapper class, namely by including a list of errors in the AccessError exception.