Combine view and update into single action
The jury's still out on whether this is too much for one action, but I
want to at least ensure we're consistent with the desired outcomes
approach and this is a bit simpler for the time being.

We'll want to revisit this pattern in future and make sure we're all
happy with it, but it does reduce some duplication (potentially at the
cost of readability).