Use new complexity level selection page
for both single-service and cohort referrals.

So that we don't have 2 differing code paths, we can use the
"cohort"-designed page and data structure for single-service referrals
too.

This also updates the '/desired-outcomes' form submit url to go to this new
one.

There's some cleanup that will need to come in future, once we're
finally able to get rid of the singular `serviceCategoryId` field on a
`DraftReferral` (after we've removed everywhere we refer to it), so in
the meantime, so as not to break the referral journey, we're setting
both in the same controller action.