Show error when there's a conflicting appointment
The Community API responds with a 409 status, which is passed on from
the Interventions Service to the UI, which we detect and then classify
as a `ServerError` here.

Before this change, the UI would the user back to the intervention
progress page, as if it had been a successful edit, even though the
Community API has rejected this update.

I initially tried using our existing formError pattern to pass in this
error, but as we're not targeting a single input field here, like we do
elsewhere (we're targeting a collection of input fields, so we can't use
the existing macros - see screenshot), I've had to handle this slightly
differently, hence the addition of the serverError argument to the
`EditSessionPresenter`.