Add `contractType` to intervention model
We'll be using this in place of the `serviceCategory`, now that we have
cohort referrals that have their own contract type - e.g. "Women's
services" could have multiple service categories (Accommodation, Social
Inclusion) but in dashboards we want to display "Women's services",
rather than a list of all the service categories.

This refactors the interventions service contract test to use the
factory-version of an intervention, so we don't have to remember to add
fields every time here.