Add service user personal details to referral check-answers
This is a replay of the data displayed on the
/referrals/:id/service-user-details page. These two pages should be
consistent.

However, this isn’t the same data as shown in the designs for the
check-your-answers page, which includes things like address and sexual
orientation. I think that we should consider it a separate piece of work
to add these fields (adding them to the service-user-details page and as
a consequence the check-answers page).