Extract business logic to a class for reuse
The piece of knowledge that a cohort intervention is one that has
multiple service categories should only exist in one place in the app.
It’s currently living in a presenter, but I need to use it in another
presenter.

If we had a "model" class called Intervention then we’d stick this as a
method in there, but we don’t - we just have a DTO interface that
contains the data returned by the interventions service API. I had
thought that perhaps we could create a model class that decorates the
DTO, allowing us to add further methods and properties. But there
doesn’t seem to be a simple way to do that in TypeScript.

So for now, I’ll just create a "decorator" class, which is really just a
utility class that takes a DTO and defines some properties derived from
it. "Decorator" is a bit of a misuse of the term; but the alternative
was another "Utils" type thing e.g. "InterventionUtils", which seemed a
bit vague. At least the naming I’ve gone for communicates what I was
_trying_ to do.