Update action plan activity page to latest design
We're now adding one activity per page and not replaying back the
previously added activities until the review page.

To try and keep this change as small as possible (I know it's a big
change already), I'm still submitting activities against a single (the
first) desired outcome to match the current API spec, but we'll want to
update the contracts for this as soon as possible and store action plan
activities by themselves without being linked to a specific desired
outcome.