Add `maximumEnforceableDays` to `draftReferral`
This will replace the current `maximumRarDays` and `usingRarDays`
fields, now that enforceable days is not an optional field - PPs always
have to specify a number of enforceable days.

I've left the old fields in there so as not to break the API at this
stage, but marked them as deprecated. We'll remove them once this change
is merged.