Require user to enter risk information on referral
This was previously optional. But, now that it's going to, for now, be
the only way for probation practitioners to communicate risk information
to the service provider, we’re making it mandatory.

We also impose a length limit, which is required by the Assess Risks and
Needs API that this data will ultimately end up in.