Replace risk info with placeholder on details page
The interventions service is going to wipe the additionalRiskInformation
field after the referral is submitted, so we won’t be able to display
the risk info using this field.

In the future, we will need to fetch this risk info from the Assess
Risks and Needs API, once they’ve built that functionality.