Remove 'additionalRiskInformation' field from sent referral DTO.
The way we have had to do this is kinda crappy - using JsonView and
having the relevant controllers set this manually. However it's the
simplest way to acheieve what we want in the short term, and it's
very testable for the existing endpoints. The worst case scenario is
that we include it, but it's always null for sent referrals anyway
so even that doesn't matter too much.