revert using user auth token for community api offender access endpoint
this commit reverts some recent changes which pass through the client supplied
access token to community api. this was never really a good idea, and it turns
out that the user type token does not have the required permissions to access the
endpoint. the implementation in this commit means that the offender access API call
always succeeds, and the implementation of the correct token config is coming in a
following commit.