comment out non-functional community-api access check until we can figure out how to get it working
the way community api requires it's access token involves customizing the token request with the
current user's username. sadly due to quirks with spring security we cannot get this flow working.
without this, the check appears to grant access regardless of the user making the request. removing
this code makes it clear that no access check is happening.