Fix logic for desired outcomes link on cohort refs
The second 'select desired outcomes' link was inaccessible, as the
previous logic looped over the `intervention`'s service categories,
rather than the referrals, meaning it was expecting a service category
not selected on the referral to have desired outcomes set against it.

Because of the way we sort service categories before looping over them,
by manually hardcoding the ids here in tests we ensure this behaviour is
properly tested.