Pass optional token to RestClient HTTP methods
We eventually want to pass a token to all these calls in the codebase,
as it allows us to pass a `RestClient` via Dependency Injection and
makes testing this a lot easier (as we can't easily mock `new
RestClient` due to the way TS modules work.

If there's no token, we make an unauthenticated request.

This also removes the unused `stream` method which we took from the
HMPPS Typescript template codebase as we've never used it (and I don't
think we plan to)