MONSTER renaming effort of everything relating to action plan appointments.
The core change here is that action plan appointments are now action plan sessions.

The rationale for this is that a subsequent change will turn what are currently 'action
plan appointments' into generic 'appointment' entities that can be used outside of action
plans. By changing all current usages of the action plan appointment to action plan
session, the hope is that the future commits will be less confusing about what is an
appointment and what is a session.

The API is unchanged - which does make for some slightly confusing endpoint names.
However the external behaviour after the whole appointment refactor should be no
different to now, so this is nothing to worry about.