only create action plan sessions (no appointments) on action plan submission.
this is quite an important change. instead of creating sessions with null appointment
details, we now just create sessions basically as placeholders which can later hold
apointments. conceptually it's better this way. it's a shame we still have to munge
the combination of session and appointment together in the DTO, but we can always
change this later.