make community api booking service totally ignorant of action plans and action plan sessions.
this changes removes all mentions of action plans from the community API booking service,
opening this class to usage by other appointment related entities. the change is a bit
messy, but mostly i think it's better. the associated changes in the action plan sessions
service are also included. which involves a little messing around with the way we save new
appointments. it's mostly fine, but will require revisiting when we allow more than one
appointment to be created against the same action plan session.