Show 'accessErrors' returned from the interventions service in the auth error page.
this commit removes the strange 'sanitizedError' class and favours the 'http-errors'
implementation which is better designed to just wrap http errors with minimal fuss.

the rest client implementation is updated to return the HttpError type from 'http-errors'
which mean that in the interventions service we don't have to wrap rest client errors
with the custom type just for validation errors. we simply utilize the fact that the
HttpError class from superagent contains the response in the error object.