Allow inline GA-variable-setting script with CSP
This allows us to set the GA uaId variable in views with a script, as we
attempted to do before in f10c703 - it was blocked due to our CSP
forbidding inline `<script>` execution. Supplying a whitelisted unique
nonce attribute to the script tag permits this execution.

See
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src
for more information.

We also allow scripts from the google analytics domain here.