Whitelist www.google-analytics.com scripts
On dev, we're getting:

```
Content Security Policy: The page's settings blocked the loading of a
resource at https://www.google-analytics.com/... ['default-src']
```

The `default-src` directive serves as a fallback for the other CSP fetch
directives[0] (e.g. `script-src`) so I've opted to add it to that directive
rather than `default-src`. I'm hoping this will fix the client-side
error.

[0]:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/default-src