Update Appointment model
We add an ID property, and make scheduling information non-nullable.

This means that an ActionPlanAppointment, which is actually a mixture of
an appointment and a session, is no longer an Appointment. (We’ll
probably update the action plan APIs to undo this mixture and make the
separate entities explicit at some point.)