Allow users to go back to EOSR creation page
if they navigated away from the page without submitting the draft End of
Service Report. If they have already started the End of Service report,
the button text will display 'continue' rather than 'create'.

As part of this, we want to ensure that we don't try and create a new
End of Service report here when clicking this button if one already
exists - we could implement this logic on the backend, so that it won't
try and create a duplicate EOSR, but this felt like an easy approach for
now.